avoid duplicated feeds in the database
it's a very naive approach. we hash feeds' location with the username (if it's set)
and then store the hash in database. this hash should be unique for each tuple
(username, host, port). We then check feed's title to differentiate between feeds
on the same page. XXX: this will probably break at some point so think of a better
way to do that. Closes #15